barry-client/pyproject.toml
[project]
name = "barry-client"
version = "0.1.0"
description = "Remote MCP client for Barry server with Gemini integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.9.0",
    "google-genai>=0.1.0",
    "python-dotenv>=1.0.0",
    "ollama>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
